clean.py script no longer removes layer.csv or material.csv or other blade setup files

+++ clean.py
@@ -5,24 +5,28 @@ import os
 import shutil
 
 
-def clean_blade_path(blade_path):
+def clean_blade_path(blade_path,
+    except_list=['blade_definition.csv',
+                 'layers.csv',
+                 'materials.csv',
+                 'airfoils',
+                 'percent_areas.csv',
+                 'percent_masses.csv',
+                 'config_name.txt',
+                 'blade_props_from_Sandia.csv',
+                 'blade_props_from_VABS.csv']):
     """Removes all files except blade defintion and airfoil path."""
     clean_list = os.listdir(blade_path)
-    clean_list.remove('blade_definition.csv')
-    try:
-        clean_list.remove('materials.csv')
-    except ValueError:
-        pass
-    clean_list.remove('airfoils')
     for item in clean_list:
-        item = os.path.join(blade_path,item)
-        if os.path.exists(item):
-            if os.path.isfile(item):
-                os.remove(item)
-                print " [Deleted file] {0}".format(item)
-            else:
-                shutil.rmtree(item)
-                print " [Deleted path] {0}".format(item)
+        if item not in except_list:
+            item = os.path.join(blade_path,item)
+            if os.path.exists(item):
+                if os.path.isfile(item):
+                    os.remove(item)
+                    print " [Deleted file] {0}".format(item)
+                else:
+                    shutil.rmtree(item)
+                    print " [Deleted path] {0}".format(item)
 
-# clean_blade_path('sandia_blade')
+clean_blade_path('sandia_blade')
 clean_blade_path('biplane_blade')
